mover algunas funciondes de TypeChecker a Types
Y actualizar una prueba
